sigmoid & cross entropy for 3rd loss function

+++ src/model.py
@@ -23,11 +23,11 @@ def get_loss(prediction, ground_truth, base_price, mask, batch_size, alpha, beta
         F.relu(pre_pw_dif * gt_pw_dif * mask_pw)
     )
     # Scale ground truth return ratios and use sigmoid to convert to probabilities
-    gt_labels = torch.where(ground_truth > 0, torch.tensor(1, device='cuda'), torch.tensor(0, device='cuda'))
+    gt_labels = torch.where(ground_truth > 0, torch.tensor(1.0, device='cuda'), torch.tensor(0.0, device='cuda'))
     # Scale predicted return ratios and use sigmoid to convert to probabilities
     pred_probs = torch.sigmoid(return_ratio*scale_factor)
     # Calculate classification loss using probabilities
-    classification_loss = F.binary_cross_entropy(pred_probs, gt_labels, reduction='mean')
+    classification_loss = F.binary_cross_entropy(pred_probs, gt_labels.float(), reduction='mean')
     loss = reg_loss + alpha * rank_loss + beta * classification_loss
     return loss, reg_loss, rank_loss, classification_loss, return_ratio
